Added fallback for when vimeo API fails
Sometimes the getVideoWidth and getVideoHeight commands don't return anything,
causing the repositionVideo() function to fail and the video to show up in the
wrong position. For those cases, this fallback will simply show the video at
a size which fills the container entirely, which is not ideal because it will
not crop letterboxing, but it's better than when the video doesn't show
properly at all.